+++ Potato Engine/Engine/Vector2.cpp
@@ -9,18 +9,18 @@ Vector2::Vector2() : x(0), y(0)
 }
 
 
-Vector2::Vector2(double x, double y) : x(x), y(y)
+Vector2::Vector2(float x, float y) : x(x), y(y)
 {
 }
 
 
-double Vector2::getLength() const
+float Vector2::getLength() const
 {
 	return sqrt(getLengthSquared());
 }
 
 
-double Vector2::getLengthSquared() const
+float Vector2::getLengthSquared() const
 {
 	return x * x + y * y;
 }
@@ -28,7 +28,7 @@ double Vector2::getLengthSquared() const
 
 Vector2 Vector2::normalized() const
 {
-	double length = getLength();
+	float length = getLength();
 	if (length == 0)
 	{
 		return *this;
@@ -39,7 +39,7 @@ Vector2 Vector2::normalized() const
 
 void Vector2::normalize()
 {
-	double length = getLength();
+	float length = getLength();
 	if (length == 0)
 	{
 		return;
@@ -49,27 +49,27 @@ void Vector2::normalize()
 }
 
 
-void Vector2::rotateCCWDegrees(double degrees)
+void Vector2::rotateCCWDegrees(float degrees)
 {
 	// To rotate a vector counter-clockwise (CCW) by an angle
 	// the following theorem is applied (for CCW rotation):
-	// x2 = x1 * cos(theta) - y1 * sin(theta)
-	// y2 = x1 * sin(theta) + y1 * cos(theta)
-	rotateCCWRadians(M_PI / 180 * degrees);
+	// x2 = x1 * cosf(theta) - y1 * sinf(theta)
+	// y2 = x1 * sinf(theta) + y1 * cosf(theta)
+	rotateCCWRadians((float)M_PI / 180 * degrees);
 }
 
 
-void Vector2::rotateCCWRadians(double radians)
+void Vector2::rotateCCWRadians(float radians)
 {
 	// To rotate a vector counter-clockwise (CCW) by an angle
 	// the following theorem is applied (for CCW rotation):
-	// x2 = x1 * cos(theta) - y1 * sin(theta)
-	// y2 = x1 * sin(theta) + y1 * cos(theta)
-	double sinRot = sin(radians);
-	double cosRot = cos(radians);
+	// x2 = x1 * cosf(theta) - y1 * sinf(theta)
+	// y2 = x1 * sinf(theta) + y1 * cosf(theta)
+	float sinRot = sinf(radians);
+	float cosRot = cosf(radians);
 
-	double oldX = x;
-	double oldY = y;
+	float oldX = x;
+	float oldY = y;
 
 	x = oldX * cosRot - oldY * sinRot;
 	y = oldX * sinRot + oldY * cosRot;
@@ -92,7 +92,7 @@ Vector2 & Vector2::operator-=(const Vector2 & rhs)
 }
 
 
-Vector2 & Vector2::operator*=(double num)
+Vector2 & Vector2::operator*=(float num)
 {
 	this->x *= num;
 	this->y *= num;
@@ -100,7 +100,7 @@ Vector2 & Vector2::operator*=(double num)
 }
 
 
-Vector2 & Vector2::operator/=(double num)
+Vector2 & Vector2::operator/=(float num)
 {
 	this->x /= num;
 	this->y /= num;
@@ -108,13 +108,13 @@ Vector2 & Vector2::operator/=(double num)
 }
 
 
-double Vector2::distance(const Vector2 & lhs, const Vector2 & rhs)
+float Vector2::distance(const Vector2 & lhs, const Vector2 & rhs)
 {
 	return Vector2(lhs.x - rhs.x, lhs.y - rhs.y).getLength();
 }
 
 
-double Vector2::dot(const Vector2 & lhs, const Vector2 & rhs)
+float Vector2::dot(const Vector2 & lhs, const Vector2 & rhs)
 {
 	return (lhs.x * rhs.x + lhs.y * rhs.y);
 }
@@ -138,25 +138,25 @@ Vector2 operator-(Vector2 lhs, const Vector2 & rhs)
 }
 
 
-Vector2 operator*(double num, const Vector2 & v2)
+Vector2 operator*(float num, const Vector2 & v2)
 {
 	return Vector2(v2.x * num, v2.y * num);
 }
 
 
-Vector2 operator/(double num, const Vector2 & v2)
+Vector2 operator/(float num, const Vector2 & v2)
 {
 	return Vector2(v2.x / num, v2.y / num);
 }
 
 
-Vector2 operator*(Vector2 v2, double num)
+Vector2 operator*(Vector2 v2, float num)
 {
 	return operator*(num, v2);
 }
 
 
-Vector2 operator/(Vector2 v2, double num)
+Vector2 operator/(Vector2 v2, float num)
 {
 	return operator/(num, v2);
 }
